feat(client): first draft of mailing
