+++ code/libs/handlebar/README.md
@@ -1,0 +1,7 @@
+# handlebar
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test handlebar` to execute the unit tests via [Jest](https://jestjs.io).
