+++ code/libs/handlebar/src/handlebar.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { HandlebarService } from './services';
+
+@Module({
+  providers: [HandlebarService],
+  exports: [HandlebarService],
+})
+export class HandlebarModule {}
