+++ code/libs/handlebar/src/index.ts
@@ -1,0 +1,2 @@
+export * from './handlebar.module';
+export * from './services';
