+++ code/libs/handlebar/src/services/handlebar/index.ts
@@ -1,0 +1,4 @@
+// start:ng42.barrel
+export * from './handlebar.service';
+// end:ng42.barrel
+
