+++ code/libs/handlebar/src/services/index.ts
@@ -1,0 +1,4 @@
+// start:ng42.barrel
+export * from './handlebar';
+// end:ng42.barrel
+
