+++ code/libs/mailing/src/index.ts
@@ -1,1 +1,2 @@
 export * from './mailing.module';
+export * from './services';
