+++ code/libs/mailing/src/mailing.module.ts
@@ -1,8 +1,9 @@
 import { Module } from '@nestjs/common';
+import { MailingService } from './services';
 
 @Module({
-  controllers: [],
-  providers: [],
-  exports: [],
+  imports: [HandlebarModule],
+  providers: [MailingService],
+  exports: [MailingService],
 })
 export class MailingModule {}
