+++ code/libs/mailing/src/services/index.ts
@@ -1,0 +1,4 @@
+// start:ng42.barrel
+export * from './mailing';
+// end:ng42.barrel
+
