+++ code/libs/mailing/src/services/mailing/index.ts
@@ -1,0 +1,4 @@
+// start:ng42.barrel
+export * from './mailing.service';
+// end:ng42.barrel
+
