+++ code/libs/mailing/src/services/mailing/mailing.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { MailingService } from './mailing.service';
+
+describe('MailingService', () => {
+  let service: MailingService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [MailingService],
+    }).compile();
+
+    service = module.get<MailingService>(MailingService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
