+++ code/libs/mailing/src/services/notification/notification.service.ts
@@ -1,0 +1,4 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class NotificationService {}
