+++ code/libs/notifications/src/index.ts
@@ -1,1 +1,2 @@
 export * from './notifications.module';
+export * from './services';
