+++ code/libs/notifications/src/notifications.module.ts
@@ -1,8 +1,10 @@
+import { MailingModule } from '@badman/mailing';
 import { Module } from '@nestjs/common';
+import { NotificationService } from './services';
 
 @Module({
-  controllers: [],
-  providers: [],
+  imports: [MailingModule],
+  providers: [NotificationService],
   exports: [],
 })
 export class NotificationsModule {}
