+++ code/libs/notifications/src/services/index.ts
@@ -1,0 +1,4 @@
+// start:ng42.barrel
+export * from './notification';
+// end:ng42.barrel
+
