+++ code/libs/notifications/src/services/notification/index.ts
@@ -1,0 +1,4 @@
+// start:ng42.barrel
+export * from './notification.service';
+// end:ng42.barrel
+
