+++ code/tsconfig.base.json
@@ -20,6 +20,7 @@
       "@badman/api/database": ["libs/api/database/src/index.ts"],
       "@badman/api/generator": ["libs/api/generator/src/index.ts"],
       "@badman/api/grapqhl": ["libs/api/grapqhl/src/index.ts"],
+      "@badman/handlebar": ["libs/handlebar/src/index.ts"],
       "@badman/health": ["libs/health/src/index.ts"],
       "@badman/mailing": ["libs/mailing/src/index.ts"],
       "@badman/notifications": ["libs/notifications/src/index.ts"],
